@@ -4,16 +4,18 @@ public abstract class CollectionSettingsBase : ISettings
 {
     //for performance reasons, all delimiters and escaped characters are single chars.
     //this makes a parsing grammar to conform LL1 rules and is very beneficial to overall parsing performance 
-    public char ListDelimiter { get; }
-    public char NullElementMarker { get; }
-    public char EscapingSequenceStart { get; }
-    public char? Start { get; }
-    public char? End { get; }
+    public char ListDelimiter { get; private set; }
+    public char NullElementMarker { get; private set; }
+    public char EscapingSequenceStart { get; private set; }
+    public char? Start { get; private set; }
+    public char? End { get; private set; }
     /// <summary>
     /// Capacity used for creating initial collection/list/array. Use no value (null) to calculate capacity each time based on input  
     /// </summary>
-    public byte? DefaultCapacity { get; }
-
+    public byte? DefaultCapacity { get; private set; }
+#if NET
+    [System.Text.Json.Serialization.JsonConstructor]
+#endif    
     protected CollectionSettingsBase(char listDelimiter, char nullElementMarker, char escapingSequenceStart, char? start, char? end, byte? defaultCapacity)
     {
         if (listDelimiter == nullElementMarker ||
@@ -55,19 +57,25 @@ Start ('{start}') and end ('{end}') can be equal to each other");
     }
 }
 
-public class CollectionSettings : CollectionSettingsBase
+public sealed class CollectionSettings : CollectionSettingsBase
 {
     public static CollectionSettings Default { get; } = new('|', '∅', '\\', null, null, null);
 
+#if NET
+    [System.Text.Json.Serialization.JsonConstructor]
+#endif 
     public CollectionSettings(char listDelimiter, char nullElementMarker, char escapingSequenceStart,
         char? start, char? end, byte? defaultCapacity)
         : base(listDelimiter, nullElementMarker, escapingSequenceStart, start, end, defaultCapacity) { }
 }
 
-public class ArraySettings : CollectionSettingsBase
+public sealed class ArraySettings : CollectionSettingsBase
 {
     public static ArraySettings Default { get; } = new('|', '∅', '\\', null, null, null);
 
+#if NET
+    [System.Text.Json.Serialization.JsonConstructor]
+#endif 
     public ArraySettings(char listDelimiter, char nullElementMarker, char escapingSequenceStart,
         char? start, char? end, byte? defaultCapacity)
         : base(listDelimiter, nullElementMarker, escapingSequenceStart, start, end, defaultCapacity) { }
@@ -76,20 +84,23 @@ public class ArraySettings : CollectionSettingsBase
 
 
 
-public class DictionarySettings : ISettings
+public sealed class DictionarySettings : ISettings
 {
-    public char DictionaryPairsDelimiter { get; }
-    public char DictionaryKeyValueDelimiter { get; }
-    public char NullElementMarker { get; }
-    public char EscapingSequenceStart { get; }
-    public char? Start { get; }
-    public char? End { get; }
-    public DictionaryBehaviour Behaviour { get; }
+    public char DictionaryPairsDelimiter { get; private set; }
+    public char DictionaryKeyValueDelimiter { get; private set; }
+    public char NullElementMarker { get; private set; }
+    public char EscapingSequenceStart { get; private set; }
+    public char? Start { get; private set; }
+    public char? End { get; private set; }
+    public DictionaryBehaviour Behaviour { get; private set; }
     /// <summary>
     /// Capacity used for creating initial dictionary. Use no value (null) to calculate capacity each time based on input  
     /// </summary>
-    public byte? DefaultCapacity { get; }
+    public byte? DefaultCapacity { get; private set; }
 
+#if NET
+    [System.Text.Json.Serialization.JsonConstructor]
+#endif 
     public DictionarySettings(char dictionaryPairsDelimiter, char dictionaryKeyValueDelimiter, char nullElementMarker, char escapingSequenceStart, char? start, char? end, DictionaryBehaviour behaviour, byte? defaultCapacity)
     {
         if (dictionaryPairsDelimiter == dictionaryKeyValueDelimiter ||
