@@ -1,17 +1,13 @@
 ﻿namespace Nemesis.TextParsers.Settings;
 
-public sealed class FactoryMethodSettings : ISettings
+#if NET
+[method: System.Text.Json.Serialization.JsonConstructor]
+#endif 
+public sealed class FactoryMethodSettings(string factoryMethodName, string emptyPropertyName, string nullPropertyName) : ISettings
 {
-    public string FactoryMethodName { get; }
-    public string EmptyPropertyName { get; }
-    public string NullPropertyName { get; }
-
-    public FactoryMethodSettings(string factoryMethodName, string emptyPropertyName, string nullPropertyName)
-    {
-        FactoryMethodName = factoryMethodName;
-        EmptyPropertyName = emptyPropertyName;
-        NullPropertyName = nullPropertyName;
-    }
+    public string FactoryMethodName { get; private set; } = factoryMethodName;
+    public string EmptyPropertyName { get; private set; } = emptyPropertyName;
+    public string NullPropertyName { get; private set; } = nullPropertyName;
 
     public static FactoryMethodSettings Default { get; } = new("FromText", "Empty", "Null");
 
